Split Unit tests to separated files based on function
