+++ Tests/Fake.DataAccess.UnitTests/Random/RandomScalarProvider/HexadecimalTests.cs
@@ -1,0 +1,115 @@
+﻿using FluentAssertions;
+using System.Collections.Generic;
+using System.Globalization;
+using System.Linq;
+using System;
+using Xunit;
+
+namespace Fake.DataAccess.UnitTests.Random.RandomScalarProvider
+{
+    public class HexadecimalTests : RandomScalarProviderTestsBase
+    {
+        [Fact]
+        public void ShouldReturnValidHexadecimal()
+        {
+            // Arrange
+            long number;
+
+            // Act
+            IEnumerable<string> values = Enumerable.Range(1, TestingSetSize).Select(_ => SUT.Hexadecimal());
+
+            // Assert
+            values.Should().OnlyContain(value => long.TryParse(value, NumberStyles.HexNumber, CultureInfo.InvariantCulture, out number));
+        }
+
+        [Fact]
+        public void ShouldReturnBetweenDefaultMinAndMax()
+        {
+            // Arrange
+            const long minValue = long.MinValue;
+            const long maxValue = long.MaxValue;
+
+            // Act
+            IEnumerable<long> values = Enumerable.Range(1, TestingSetSize)
+                .Select(_ => SUT.Hexadecimal())
+                .Select(value => Convert.ToInt64(value, 16));
+
+            // Assert
+            values.Should().OnlyContain(value => value >= minValue && value <= maxValue);
+        }
+
+        [Fact]
+        public void ShouldReturnHigherThanMin()
+        {
+            // Arrange
+            long minValue = RandomNumber(long.MinValue, long.MaxValue);
+
+            // Act
+            IEnumerable<long> values = Enumerable.Range(1, TestingSetSize)
+                .Select(_ => SUT.Hexadecimal(min: minValue))
+                .Select(value => Convert.ToInt64(value, 16));
+
+            // Assert
+            values.Should().OnlyContain(value => value >= minValue);
+        }
+
+        [Fact]
+        public void ShouldReturnLowerThanMax()
+        {
+            // Arrange
+            long maxValue = RandomNumber(long.MinValue, long.MaxValue);
+
+            // Act
+            IEnumerable<long> values = Enumerable.Range(1, TestingSetSize)
+                .Select(_ => SUT.Hexadecimal(max: maxValue))
+                .Select(value => Convert.ToInt64(value, 16));
+
+            // Assert
+            values.Should().OnlyContain(value => value <= maxValue);
+        }
+
+        [Fact]
+        public void ShouldFailForInvertedMinMax()
+        {
+            // Arrange
+            const long minValue = long.MaxValue;
+            const long maxValue = long.MinValue;
+
+            // Act
+            Action value = () => SUT.Hexadecimal(min: minValue, max: maxValue);
+
+            // Assert
+            value.Should().Throw<ArgumentOutOfRangeException>();
+        }
+
+        [Fact]
+        public void ShouldWorkForMinBoundaryCondition()
+        {
+            // Arrange
+            const long minValue = long.MinValue;
+
+            // Act
+            IEnumerable<long> values = Enumerable.Range(1, TestingSetSize)
+                .Select(_ => SUT.Hexadecimal(min: minValue, max: minValue + 1))
+                .Select(value => Convert.ToInt64(value, 16));
+
+            // Assert
+            values.Should().Contain(minValue);
+        }
+
+        [Fact]
+        public void ShouldWorkForMaxBoundaryCondition()
+        {
+            // Arrange
+            const long maxValue = long.MaxValue;
+
+            // Act
+            IEnumerable<long> values = Enumerable.Range(1, TestingSetSize)
+                .Select(_ => SUT.Hexadecimal(min: maxValue - 1, max: maxValue))
+                .Select(value => Convert.ToInt64(value, 16));
+
+            // Assert
+            values.Should().Contain(maxValue);
+        }
+    }
+}
