+++ Tests/Fake.DataAccess.UnitTests/Random/RandomScalarProvider/StringsTests.cs
@@ -1,0 +1,231 @@
+﻿using FluentAssertions;
+using System.Collections.Generic;
+using System.Linq;
+using Xunit;
+
+namespace Fake.DataAccess.UnitTests.Random.RandomScalarProvider
+{
+    public class StringsTests : RandomScalarProviderTestsBase
+    {
+        [Fact]
+        public void ShouldReturnCorrectCount()
+        {
+            // Act
+            IEnumerable<string> values = SUT.Strings(TestingSetSize);
+
+            // Assert
+            values.Should().HaveCount(TestingSetSize);
+        }
+
+        [Fact]
+        public void ShouldReturnBetweenDefaultMinAndMaxLength()
+        {
+            // Arrange
+            const short minValue = 1;
+            const short maxValue = short.MaxValue;
+
+            // Act
+            IEnumerable<string> values = SUT.Strings(TestingSetSize);
+
+            // Assert
+            values.Should().OnlyContain(value => value.Length >= minValue && value.Length <= maxValue);
+        }
+
+        [Fact]
+        public void ShouldReturnBetweenDefaultMinAndMaxChar()
+        {
+            // Arrange
+            const char minValue = char.MinValue;
+            const char maxValue = char.MaxValue;
+
+            // Act
+            IEnumerable<string> values = SUT.Strings(1, minLength: TestingSetSize, maxLength: TestingSetSize);
+
+            // Assert
+            values.First().ToCharArray().Should().OnlyContain(@char => @char >= minValue && @char <= maxValue);
+        }
+
+        [Fact]
+        public void ShouldReturnHigherThanMinLength()
+        {
+            // Arrange
+            short minValue = (short)RandomNumber(1, short.MaxValue);
+
+            // Act
+            IEnumerable<string> values = SUT.Strings(TestingSetSize, minLength: minValue);
+
+            // Assert
+            values.Should().OnlyContain(value => value.Length >= minValue);
+        }
+
+        [Fact]
+        public void ShouldReturnLowerThanMaxLength()
+        {
+            // Arrange
+            short maxValue = (short)RandomNumber(1, short.MaxValue);
+
+            // Act
+            IEnumerable<string> values = SUT.Strings(TestingSetSize, maxLength: maxValue);
+
+            // Assert
+            values.Should().OnlyContain(value => value.Length <= maxValue);
+        }
+
+        [Fact]
+        public void ShouldReturnHigherThanMinChar()
+        {
+            // Arrange
+            char minValue = (char)RandomNumber(1, char.MaxValue);
+
+            // Act
+            IEnumerable<string> values = SUT.Strings(1, minLength: TestingSetSize, maxLength: TestingSetSize, minChar: minValue);
+
+            // Assert
+            values.First().ToCharArray().Should().OnlyContain(@char => @char >= minValue);
+        }
+
+        [Fact]
+        public void ShouldReturnLowerThanMaxChar()
+        {
+            // Arrange
+            char maxValue = (char)RandomNumber(1, char.MaxValue);
+
+            // Act
+            IEnumerable<string> values = SUT.Strings(TestingSetSize, minLength: TestingSetSize, maxLength: TestingSetSize, maxChar: maxValue);
+
+            // Assert
+            values.First().ToCharArray().Should().OnlyContain(@char => @char <= maxValue);
+        }
+
+        [Fact]
+        public void ShouldReturnZeroLength()
+        {
+            // Arrange
+            const short length = 0;
+
+            // Act
+            IEnumerable<string> values = SUT.Strings(1, minLength: length, maxLength: length);
+
+            // Assert
+            values.Should().OnlyContain(value => string.IsNullOrEmpty(value));
+        }
+
+        [Fact]
+        public void ShouldReturnEmptyEnumerationForNegativeCount()
+        {
+            // Arrange
+            short count = (short)RandomNumber(short.MinValue, 0);
+
+            // Act
+            IEnumerable<string> values = SUT.Strings(count);
+
+            // Assert
+            values.Should().NotBeNull();
+            values.Should().BeEmpty();
+        }
+
+        [Fact]
+        public void ShouldReturnEmptyEnumerationForInvertedMinMaxLength()
+        {
+            // Arrange
+            const short minValue = short.MaxValue;
+            const short maxValue = 1;
+
+            // Act
+            IEnumerable<string> values = SUT.Strings(TestingSetSize, minLength: minValue, maxLength: maxValue);
+
+            // Assert
+            values.Should().NotBeNull();
+            values.Should().BeEmpty();
+        }
+
+        [Fact]
+        public void ShouldReturnEmptyEnumerationForInvertedMinMaxChar()
+        {
+            // Arrange
+            const char minValue = char.MaxValue;
+            const char maxValue = char.MinValue;
+
+            // Act
+            IEnumerable<string> values = SUT.Strings(TestingSetSize, minChar: minValue, maxChar: maxValue);
+
+            // Assert
+            values.Should().NotBeNull();
+            values.Should().BeEmpty();
+        }
+
+        [Fact]
+        public void ShouldReturnEmptyEnumerationForNegativeMinLength()
+        {
+            // Arrange
+            short minValue = (short)RandomNumber(short.MinValue, -1);
+
+            // Act
+            IEnumerable<string> values = SUT.Strings(TestingSetSize, minLength: minValue);
+
+            // Assert
+            values.Should().NotBeNull();
+            values.Should().BeEmpty();
+        }
+
+        [Theory]
+        [InlineData(0)]
+        public void ShouldWorkForMinLengthBoundaryCondition(short boundaryValue)
+        {
+            // Act
+            IEnumerable<string> values = SUT.Strings(TestingSetSize, minLength: boundaryValue, maxLength: (short)(boundaryValue + 1));
+
+            // Assert
+            values.Should().Contain(value => value.Length >= boundaryValue);
+        }
+
+        [Theory]
+        [InlineData(short.MaxValue)]
+        public void ShouldWorkForMaxLengthBoundaryCondition(short boundaryValue)
+        {
+            // Act
+            IEnumerable<string> values = SUT.Strings(TestingSetSize, minLength: (short)(boundaryValue - 1), maxLength: boundaryValue);
+
+            // Assert
+            values.Should().Contain(value => value.Length >= boundaryValue);
+        }
+
+        [Fact]
+        public void ShouldWorkForMinCharBoundaryCondition()
+        {
+            // Arrange
+            const char minValue = char.MinValue;
+
+            // Act
+            IEnumerable<char> values = SUT.Strings(1, TestingSetSize, TestingSetSize, minChar: minValue, maxChar: (char)(minValue + 1)).First().ToCharArray();
+
+            // Assert
+            values.Should().Contain(minValue);
+        }
+
+        [Fact]
+        public void ShouldWorkForMaxCharBoundaryCondition()
+        {
+            // Arrange
+            const char maxValue = char.MaxValue;
+
+            // Act
+            IEnumerable<char> values = SUT.Strings(1, TestingSetSize, TestingSetSize, minChar: (char)(maxValue - 1), maxChar: maxValue).First().ToCharArray();
+
+            // Assert
+            values.Should().Contain(maxValue);
+        }
+
+        [Theory]
+        [InlineData(1)]
+        [InlineData(short.MaxValue)]
+        public void ShouldWorkForCountBoundaryCondition(short boundaryValue)
+        {
+            // Act
+            IEnumerable<string> values = SUT.Strings(boundaryValue, 1, 1);
+
+            // Assert
+            values.Should().HaveCount(boundaryValue);
+        }
+    }
+}
